+++ keepachangelog/_markdown.py
@@ -5,11 +5,11 @@ link_pattern = re.compile(r"^\[(.*)\]: (.*)\n$")
 
 
 def is_heading(line: str, heading_level: int) -> bool:
-    return line.lstrip().startswith(f"{'#' * heading_level} ")
+    return line.lstrip(" ").startswith(f"{'#' * heading_level} ")
 
 
 def from_heading(line: str, heading_level: int) -> str:
-    return line.lstrip()[heading_level:].strip(" \n")
+    return line.lstrip(" ")[heading_level:].strip(" \n")
 
 
 def unlink(value: str) -> str:
@@ -17,4 +17,18 @@ def unlink(value: str) -> str:
 
 
 def is_link(line: str) -> bool:
-    return link_pattern.fullmatch(line.lstrip()) is not None
+    return link_pattern.fullmatch(line.lstrip(" ")) is not None
+
+
+def unlist(line: str) -> str:
+    line = line.lstrip(" ")
+    if line and line[0] in "-*+":
+        list_line = line[1:]
+        # Empty list item
+        if not list_line:
+            return ""
+        # List item
+        if list_line[0] == " ":
+            return list_line[1:]
+    # Non list item (can start with a list identifier)
+    return line